Prevent context menu from closing when reference point dialog is open
Update database schema to use REAL for latitude and longitude in reference points table. Modify `useEffect` in `operational-grid.tsx` to prevent the context menu from closing when any dialog is open, specifically addressing an issue where clicking within the reference point dialog would trigger the `onClose` handler.

+++ server/db-init.ts
@@ -428,8 +428,8 @@ export async function ensureDbSchema() {
       CREATE TABLE IF NOT EXISTS reference_points (
         id SERIAL PRIMARY KEY,
         name TEXT NOT NULL,
-        latitude DOUBLE PRECISION NOT NULL,
-        longitude DOUBLE PRECISION NOT NULL,
+        latitude REAL NOT NULL,
+        longitude REAL NOT NULL,
         radius_meters INTEGER NOT NULL DEFAULT 500,
         color TEXT NOT NULL DEFAULT '#6366f1',
         created_at TIMESTAMP DEFAULT NOW()
